Moved flag_290_case into stage_dataframe module
