+++ cons_results/imputation/post_imputation.py
@@ -236,56 +236,6 @@ def rescale_imputed_values(
     return df  # Return the modified DataFrame
 
 
-def flag_290_case(
-    df: pd.DataFrame,
-    period: str,
-    reference: str,
-    question_no: str,
-    adjusted_response: str,
-) -> pd.DataFrame:
-
-    """
-    Function to flag cases for imputation where value for question is 290
-    is given with no other components
-    """
-
-    # Group and sum adjusted responses for question 290
-    question_290_df = (
-        df[df[question_no] == 290].groupby([period, reference])[adjusted_response].sum()
-    )
-
-    # Group and sum adjusted responses for all other questions
-    other_questions_df = (
-        df[df[question_no] != 290].groupby([period, reference])[adjusted_response].sum()
-    )
-
-    # Merge groupings
-    df_joined = pd.merge(
-        question_290_df,
-        other_questions_df,
-        on=[period, reference],
-    )
-
-    # Create index of pairs of period and reference numbers which need to be
-    # flagged as the special 290 case
-    flagged_pairs = df_joined[
-        (df_joined[f"{adjusted_response}_x"] != df_joined[f"{adjusted_response}_y"])
-        & (df_joined[f"{adjusted_response}_y"] == 0)
-    ].index
-
-    # Initialise flag
-    df["290_flag"] = False
-
-    # Set flag based on index
-    df.loc[
-        pd.MultiIndex.from_frame(df[[period, reference]]).isin(flagged_pairs),
-        ["290_flag"],
-    ] = True
-
-    # Return modified DataFrame
-    return df
-
-
 def rescale_290_case(
     df: pd.DataFrame,
     period: str,
