+++ tests/imputation/test_post_imputation.py
@@ -2,7 +2,6 @@ import pandas as pd
 import pytest
 
 from cons_results.imputation.post_imputation import (
-    flag_290_case,
     rescale_290_case,
     rescale_imputed_values,
 )
@@ -85,18 +84,6 @@ def test_rescale_imputed_values_double_call(reference):
 
     print(actual_output, expected_output)
     pd.testing.assert_frame_equal(actual_output, expected_output)
-
-
-def test_flag_290_case():
-    expected_output_df = pd.read_csv("tests/data/imputation/test_data_290_cases.csv")
-
-    input_df = expected_output_df.drop(columns=["290_flag"])
-
-    output_df = flag_290_case(
-        input_df, "period", "reference", "question_no", "adjustedresponse"
-    )
-
-    pd.testing.assert_frame_equal(output_df, expected_output_df)
 
 
 def test_rescale_290_case():
